src/python/requirements.txt
fastapi
uvicorn
sentence-transformers
faiss-cpu
